+++ src/snowflake/cli/plugins/object/stage/commands.py
@@ -3,10 +3,15 @@ from __future__ import annotations
 import itertools
 from os import path
 from pathlib import Path
+from typing import List, Optional
 
 import click
 import typer
-from snowflake.cli.api.commands.flags import PatternOption
+from snowflake.cli.api.commands.flags import (
+    OnErrorOption,
+    PatternOption,
+    VariablesOption,
+)
 from snowflake.cli.api.commands.snow_typer import SnowTyper
 from snowflake.cli.api.console import cli_console
 from snowflake.cli.api.output.types import (
@@ -18,7 +23,7 @@ from snowflake.cli.api.output.types import (
 )
 from snowflake.cli.api.utils.path_utils import is_stage_path
 from snowflake.cli.plugins.object.stage.diff import DiffResult
-from snowflake.cli.plugins.object.stage.manager import StageManager
+from snowflake.cli.plugins.object.stage.manager import OnErrorType, StageManager
 
 app = SnowTyper(
     name="stage",
@@ -127,6 +132,27 @@ def stage_diff(
     """
     diff: DiffResult = stage_diff(Path(folder_name), stage_name)
     return ObjectResult(str(diff))
+
+
+@app.command("execute", requires_connection=True)
+def execute(
+    stage_path: str = typer.Argument(
+        ...,
+        help="Stage path with files to be execute. For example `@stage/dev/*`.",
+        show_default=False,
+    ),
+    on_error: OnErrorType = OnErrorOption,
+    variables: Optional[List[str]] = VariablesOption,
+    **options,
+):
+    """
+    Execute immediate all files from the stage path. Files can be filtered with glob like pattern,
+    e.g. `@stage/*.sql`, `@stage/dev/*`. Only files with `.sql` extension will be executed.
+    """
+    results = StageManager().execute(
+        stage_path=stage_path, on_error=on_error, variables=variables
+    )
+    return CollectionResult(results)
 
 
 def _get(recursive: bool, source_path: str, destination_path: str, parallel: int):